remove base filter modify filters and url move to case service
