+++ src/main/java/ca/on/oicr/gsi/dimsum/service/CaseService.java
@@ -35,6 +35,10 @@ import ca.on.oicr.gsi.dimsum.service.filtering.SampleSort;
 import ca.on.oicr.gsi.dimsum.service.filtering.TableData;
 import io.micrometer.core.instrument.Gauge;
 import io.micrometer.core.instrument.MeterRegistry;
+import ca.on.oicr.gsi.dimsum.data.Run;
+import ca.on.oicr.gsi.dimsum.service.filtering.RunFilter;
+import ca.on.oicr.gsi.dimsum.service.filtering.RunFilterKey;
+import ca.on.oicr.gsi.dimsum.service.filtering.RunSort;
 
 @Service
 public class CaseService {
@@ -208,6 +212,24 @@ public class CaseService {
     return data;
   }
 
+  public TableData<Run> getRuns(int pageSize, int pageNumber, RunSort sort, boolean descending,
+      Collection<RunFilter> filters) {
+    List<Run> baseRuns = runListManager.getRuns();
+    Stream<Run> stream = applyFiltersOnRuns(baseRuns, filters);
+    if (sort == null) {
+      sort = RunSort.COMPLETION_DATE;
+      descending = true;
+    }
+    stream = stream.sorted(descending ? sort.comparator().reversed() : sort.comparator());
+    List<Run> filteredRuns =
+        stream.skip(pageSize * (pageNumber - 1)).limit(pageSize).collect(Collectors.toList());
+    TableData<Run> data = new TableData<>();
+    data.setTotalCount(baseRuns.size());
+    data.setFilteredCount(applyFiltersOnRuns(baseRuns, filters).count());
+    data.setItems(filteredRuns);
+    return data;
+  }
+
   private Stream<Case> applyFilters(List<Case> cases, Collection<CaseFilter> filters) {
     Stream<Case> stream = cases.stream();
     if (filters != null && !filters.isEmpty()) {
@@ -221,6 +243,25 @@ public class CaseService {
         }
       }
       for (Predicate<Case> predicate : filterMap.values()) {
+        stream = stream.filter(predicate);
+      }
+    }
+    return stream;
+  }
+
+  private Stream<Run> applyFiltersOnRuns(List<Run> runs, Collection<RunFilter> filters) {
+    Stream<Run> stream = runs.stream();
+    if (filters != null && !filters.isEmpty()) {
+      Map<RunFilterKey, Predicate<Run>> filterMap = new HashMap<>();
+      for (RunFilter filter : filters) {
+        RunFilterKey key = filter.getKey();
+        if (filterMap.containsKey(key)) {
+          filterMap.put(key, filterMap.get(key).or(filter.predicate()));
+        } else {
+          filterMap.put(key, filter.predicate());
+        }
+      }
+      for (Predicate<Run> predicate : filterMap.values()) {
         stream = stream.filter(predicate);
       }
     }
