+++ src/main/resources/templates/layout.html
@@ -49,7 +49,7 @@
     </div>
     <nav th:unless="${hideNav}" class="text-right">
       <a th:href="@{/runs}" class="text-green-200 font-bold hover:underline">Runs</a>
-      <span class="text-green-200 font-bold"> | </span>
+      <span class="text-grey-200 font-bold"> | </span>
       <a th:href="@{/notifications}" class="text-green-200 font-bold hover:underline">Notifications</a>
     </nav>
   </div>
